+++ breakout-clone/game/C_Paddle.h
@@ -6,9 +6,9 @@
 class C_Paddle
 {
     private:
-    double speed = 300;// movement speed paddle
-    double width = 70;
-    double height = 10;
+    float speed = 300;// movement speed paddle
+    float width = 70;
+    float height = 10;
 
     bool move_left = false;
     bool move_right = false;
